Kafka+SparkStreaming+Docker Wordcount/requirements.txt
pyspark==2.4.7
kafka-python
findspark
